chore(eslint): upgrade to v9
