+++ package.json
@@ -12,13 +12,13 @@
   "devDependencies": {
     "@endo/eslint-plugin": "^2.2.3",
     "@google-cloud/monitoring": "^4.1.0",
-    "@jessie.js/eslint-plugin": "^0.4.1",
+    "@jessie.js/eslint-plugin": "^0.4.2",
     "@types/express": "^4.17.17",
     "@types/node": "^22.9.0",
     "ava": "^5.3.0",
     "c8": "^10.1.2",
     "conventional-changelog-conventionalcommits": "^4.6.0",
-    "eslint": "^8.57.1",
+    "eslint": "^9.0.0",
     "eslint-config-airbnb-base": "^15.0.0",
     "eslint-config-jessie": "^0.0.6",
     "eslint-config-prettier": "^9.1.0",
@@ -59,8 +59,8 @@
     "lerna": "lerna",
     "link-cli": "yarn run create-agoric-cli",
     "create-agoric-cli": "node ./scripts/create-agoric-cli.cjs",
-    "format": "yarn prettier --write .github golang packages scripts a3p-integration multichain-testing",
-    "lint:format": "yarn prettier --check .github golang packages scripts a3p-integration multichain-testing",
+    "format": "yarn prettier --write .github golang packages scripts a3p-integration multichain-testing eslint.config.mjs",
+    "lint:format": "yarn prettier --check .github golang packages scripts a3p-integration multichain-testing eslint.config.mjs",
     "lint-fix": "yarn lerna run --no-bail lint-fix",
     "lint": "run-s --continue-on-error lint:*",
     "lint:packages": "yarn lerna run --no-bail lint",
