+++ packages/cosmic-swingset/tsconfig.json
@@ -8,5 +8,6 @@
     "calc-*.js",
     "src/**/*.js",
     "test/**/*.js",
+    "*.cjs",
   ],
 }
