+++ packages/eslint-config/package.json
@@ -26,9 +26,9 @@
   ],
   "peerDependencies": {
     "@endo/eslint-plugin": "^2.2.3",
-    "@jessie.js/eslint-plugin": "^0.4.1",
+    "@jessie.js/eslint-plugin": "^0.4.2",
     "typescript-eslint": "^8.17.0",
-    "eslint": "^8.57.0",
+    "eslint": "^9.0.0",
     "eslint-config-airbnb-base": "^15.0.0",
     "eslint-plugin-github": "^5.1.4",
     "eslint-config-jessie": "^0.0.6",
